docker-app/requirements.txt
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.10.0
boto3==1.18.65
boto3-stubs==1.20.26
botocore==1.21.65
botocore-stubs==1.23.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.12
django-allauth==0.44.0
django-auditlog==1.0a1
django-axes==5.28.0
django-bootstrap4==3.0.1
django-classy-tags==3.0.1
django-common-helpers==0.9.2
django-cron==0.5.0
django-currentuser==0.5.3
django-filter==21.1
django-invitations==1.9.3
django-ipware==4.0.2
django-jsonfield==1.4.1
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-storages==1.11.1
django-tables2==2.4.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
docker==4.2.2
drf-yasg==1.20.0
Fiona==1.8.20
ghp-import==2.0.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
JSON-log-formatter==0.5.0
jsonfield==3.1.0
jsonschema==3.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.3
munch==2.5.0
mypy-boto3-s3==1.20.17
oauthlib==3.1.1
packaging==21.3
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4
pyyaml-env-tag==0.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sentry-sdk==1.5.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
swapper==1.3.0
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
watchdog==2.1.6
websocket-client==1.2.3
wrapt==1.13.3
zipp==3.6.0
